inferra-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
PyPDF2
sentence-transformers
chromadb
